+++ src/components/Feeds/Feed/Feed.jsx
@@ -1,7 +1,8 @@
 import theme from 'styles/theme';
 
-import { FeedWrapper, FeedImageArea } from './Feed.style';
+import { FeedWrapper, FeedImageArea } from './Feed.styled';
 import FeedButtonsArea from './FeedItem/ButtonsArea';
+import CommentsArea from './FeedItem/CommentsArea';
 import FeedHeader from './FeedItem/Header';
 import LikedArea from './FeedItem/LikedArea';
 
@@ -17,6 +18,7 @@ const Feed = ({ userId }) => {
       </FeedImageArea>
       <FeedButtonsArea />
       <LikedArea likeCount={50000} />
+      <CommentsArea />
     </FeedWrapper>
   );
 };
